eval/requirements.txt
openai>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
requests>=2.32.0
python-dotenv>=1.0.0
cadquery>=2.4.0
trimesh>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
tabulate>=0.9.0
